+++ grid-client/deployer/deployment_deployer.go
@@ -61,16 +61,16 @@ func (d *DeploymentDeployer) Validate(ctx context.Context, dls []*workloads.Depl
 func (d *DeploymentDeployer) GenerateVersionlessDeployments(ctx context.Context, dls []*workloads.Deployment) (map[uint32][]gridtypes.Deployment, error) {
 	gridDlsPerNodes := make(map[uint32][]gridtypes.Deployment)
 
-	err := d.assignPrivateIPs(ctx, dls)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to assign node ips")
-	}
-
 	var wg sync.WaitGroup
 	var mu sync.Mutex
 	var errs error
 
-	for _, dl := range dls {
+	newDls, err := d.assignPrivateIPs(ctx, dls)
+	if err != nil {
+		errs = multierror.Append(errs, errors.Wrap(err, "failed to assign node ips"))
+	}
+
+	for _, dl := range newDls {
 		wg.Add(1)
 		go func(dl *workloads.Deployment) {
 			defer wg.Done()
@@ -96,16 +96,14 @@ func (d *DeploymentDeployer) GenerateVersionlessDeployments(ctx context.Context,
 				newDl.Workloads = append(newDl.Workloads, qsfsWorkload)
 			}
 
+			mu.Lock()
+			defer mu.Unlock()
 			newDl.Metadata, err = dl.GenerateMetadata()
 			if err != nil {
-				mu.Lock()
-				defer mu.Unlock()
 				errs = multierror.Append(errs, errors.Wrapf(err, "failed to generate deployment '%s' metadata", dl.Name))
 				return
 			}
 
-			mu.Lock()
-			defer mu.Unlock()
 			gridDlsPerNodes[dl.NodeID] = append(gridDlsPerNodes[dl.NodeID], newDl)
 		}(dl)
 	}
@@ -172,7 +170,7 @@ func (d *DeploymentDeployer) BatchDeploy(ctx context.Context, dls []*workloads.D
 	// error is not returned immediately before updating state because of untracked failed deployments
 	for _, dl := range dls {
 		if err := d.updateStateFromDeployments(ctx, dl, newDls); err != nil {
-			return errors.Wrapf(err, "failed to update deployment '%s' state", dl.Name)
+			multiErr = multierror.Append(multiErr, errors.Wrapf(err, "failed to update deployment '%s' state", dl.Name))
 		}
 	}
 
@@ -304,18 +302,20 @@ func (d *DeploymentDeployer) calculateNetworksUsedIPs(ctx context.Context, dls [
 	return usedHosts, errs
 }
 
-func (d *DeploymentDeployer) assignPrivateIPs(ctx context.Context, dls []*workloads.Deployment) error {
+func (d *DeploymentDeployer) assignPrivateIPs(ctx context.Context, dls []*workloads.Deployment) ([]*workloads.Deployment, error) {
+	var newdls []*workloads.Deployment
 	var wg sync.WaitGroup
 	var mu sync.Mutex
 	var errs error
 
 	usedHosts, err := d.calculateNetworksUsedIPs(ctx, dls)
 	if err != nil {
-		return errors.Wrap(err, "couldn't calculate network used ips")
+		errs = multierror.Append(errs, errors.Wrap(err, "couldn't calculate networks used ips"))
 	}
 
 	for _, dl := range dls {
 		if len(dl.Vms) == 0 {
+			newdls = append(newdls, dl)
 			continue
 		}
 
@@ -392,11 +392,15 @@ func (d *DeploymentDeployer) assignPrivateIPs(ctx context.Context, dls []*worklo
 			}
 
 			dl.IPrange = ipRange
+
+			mu.Lock()
+			newdls = append(newdls, dl)
+			mu.Unlock()
 		}(dl)
 	}
 
 	wg.Wait()
-	return errs
+	return newdls, errs
 }
 
 func (d *DeploymentDeployer) syncContract(ctx context.Context, dl *workloads.Deployment) error {