+++ grid-proxy/pkg/types/nodes.go
@@ -47,7 +47,10 @@ type Node struct {
 	NumGPU            int          `json:"num_gpu" sort:"num_gpu"`
 	ExtraFee          uint64       `json:"extraFee" sort:"extra_fee"`
 	Healthy           bool         `json:"healthy"`
+	Dmi               Dmi          `json:"dmi"`
+	Speed             Speed        `json:"speed"`
 	PriceUsd          float64      `json:"price_usd" sort:"price_usd"`
+	_                 string       `sort:"free_cru"`
 }
 
 // CapacityResult is the NodeData capacity results to unmarshal json in it
@@ -86,6 +89,8 @@ type NodeWithNestedCapacity struct {
 	NumGPU            int            `json:"num_gpu"`
 	ExtraFee          uint64         `json:"extraFee"`
 	Healthy           bool           `json:"healthy"`
+	Dmi               Dmi            `json:"dmi"`
+	Speed             Speed          `json:"speed"`
 	PriceUsd          float64        `json:"price_usd"`
 }
 
@@ -139,6 +144,7 @@ type NodeFilter struct {
 	TwinID            *uint64  `schema:"twin_id,omitempty"`
 	CertificationType *string  `schema:"certification_type,omitempty"`
 	HasGPU            *bool    `schema:"has_gpu,omitempty"`
+	NumGPU            *uint64  `schema:"num_gpu,omitempty"`
 	GpuDeviceID       *string  `schema:"gpu_device_id,omitempty"`
 	GpuDeviceName     *string  `schema:"gpu_device_name,omitempty"`
 	GpuVendorID       *string  `schema:"gpu_vendor_id,omitempty"`
@@ -148,19 +154,5 @@ type NodeFilter struct {
 	PriceMin          *float64 `schema:"price_min,omitempty"`
 	PriceMax          *float64 `schema:"price_max,omitempty"`
 	Excluded          []uint64 `schema:"excluded,omitempty"`
-}
-
-// NodeGPU holds the info about gpu card
-type NodeGPU struct {
-	NodeTwinID uint32 `json:"node_twin_id"`
-	ID         string `json:"id"`
-	Vendor     string `json:"vendor"`
-	Device     string `json:"device"`
-	Contract   int    `json:"contract"`
-}
-
-// HeathReport holds the info of node health
-type HealthReport struct {
-	NodeTwinId uint32
-	Healthy    bool
+	HasIpv6           *bool    `schema:"has_ipv6,omitempty"`
 }